refactor: merge jobs and job-detail folders into single jobs folder
