+++ backend/server.js
@@ -21,8 +21,8 @@ const jobsRoutes = require("./jobs/route");
 app.use(jobsRoutes);
 
 
-const jobDetailRoutes = require("./job-detail/route");
-app.use('/api/job', jobDetailRoutes);
+// const jobDetailRoutes = require("./job-detail/route");
+// app.use('/api/job', jobDetailRoutes);
 
 app.use((err, req, res, next) => {
   console.error('Error:', err);
